requirements.txt
torch==2.2.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
pyyaml==6.0.1
requests==2.31.0
ruff==0.3.2
pytest==8.1.1
